ref: extract input from map
